Update team view

+++ he-f1/src/main/java/ch/hearc/hef1/model/CarPiece.java
@@ -33,10 +33,10 @@ public class CarPiece {
 	private RepairUpgrade repairUpgrade;
 
 	@Column
-	private double wear;
+	private int wear;
 
 	@Column
-	private double level;
+	private int level;
 
 	/**
 	 * Default constructor
@@ -53,7 +53,7 @@ public class CarPiece {
 	 * @param wear
 	 * @param level
 	 */
-	public CarPiece(int id, Piece piece, Car car, RepairUpgrade repairUpgrade, double wear, double level) {
+	public CarPiece(int id, Piece piece, Car car, RepairUpgrade repairUpgrade, int wear, int level) {
 		this.id = id;
 		this.piece = piece;
 		this.car = car;
@@ -72,7 +72,7 @@ public class CarPiece {
 	 * @param wear
 	 * @param level
 	 */
-	public CarPiece(Piece piece, Car car, RepairUpgrade repairUpgrade, double wear, double level) {
+	public CarPiece(Piece piece, Car car, RepairUpgrade repairUpgrade, int wear, int level) {
 		this.piece = piece;
 		this.car = car;
 		this.repairUpgrade = repairUpgrade;
@@ -130,7 +130,7 @@ public class CarPiece {
 		return wear;
 	}
 
-	public void setWear(double wear) {
+	public void setWear(int wear) {
 		this.wear = wear;
 	}
 
@@ -138,7 +138,7 @@ public class CarPiece {
 		return level;
 	}
 
-	public void setLevel(double level) {
+	public void setLevel(int level) {
 		this.level = level;
 	}
